+++ fast-ftrl-proximal/src/main/java/io/scaledml/core/inputformats/VowpalWabbitFormat.java
@@ -3,9 +3,9 @@ package io.scaledml.core.inputformats;
 import com.google.common.base.CharMatcher;
 import com.google.inject.Inject;
 import io.scaledml.core.SparseItem;
-import io.scaledml.ftrl.featuresprocessors.FeaturesProcessor;
 import io.scaledml.core.util.LineBytesBuffer;
 import io.scaledml.core.util.Util;
+import io.scaledml.ftrl.featuresprocessors.FeaturesProcessor;
 
 public class VowpalWabbitFormat implements InputFormat {
     private static final char NAME_CHAR = 'z';
